refactor(core): move the control-plane address into mxb-core
The address and its debug override were private to paintsync, and both binaries
need them — the studio's track generation talks to the same worker. They move to
`mxb_core::names`, which is where small shared facts with no better home go.

paintsync re-exports them, so the seven modules that already say
`paintsync::control_plane` are untouched. The address is a paint-sync fact as
much as anyone else's, and moving the call sites would have been churn for its
own sake.

Verified: cargo check clean on host and x86_64-pc-windows-gnu; 1295 + 184 tests
pass, unchanged.
